+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -24,6 +24,7 @@
 - Added Incomingcall event to support incoming call notification for Teams multipersona users
 - Added StartRecordingFailed event to indicate when the start recording API is unable to initiate the recording.
 - DTMF broadcast functionality is now supported in the Create Call, Answer Call, and Create Group Call APIs.
+- Added marker audio for streaming data.
 
 ### Breaking Changes